+++ web/usuaris/src/__tests__/organisms/FiltersWidget.test.tsx
@@ -1,9 +1,11 @@
 import { fireEvent, render, screen, waitFor } from '../test-utils'
 import { FiltersWidget } from '../../components/organisms'
 
+const mockHandleFilters = vi.fn()
+
 describe('FiltersWidget', () => {
   it('renders correctly', async () => {
-    render(<FiltersWidget />)
+    render(<FiltersWidget filters={{}} handleFilters={mockHandleFilters} />)
 
     await waitFor(() =>
       expect(screen.getByText(/especialitat/i)).toBeInTheDocument()
@@ -11,8 +13,8 @@ describe('FiltersWidget', () => {
     expect(screen.getByAltText(/calendar/i)).toBeInTheDocument()
   })
 
-  it('shows itinerary selected in dropdown', async () => {
-    render(<FiltersWidget />)
+  it('shows itinerary selected in dropdown and sends it to parent', async () => {
+    render(<FiltersWidget filters={{}} handleFilters={mockHandleFilters} />)
 
     await waitFor(() =>
       expect(screen.getByText(/especialitat/i)).toBeInTheDocument()
@@ -31,5 +33,48 @@ describe('FiltersWidget', () => {
 
     expect(dropdownHeader).toHaveTextContent('Frontend React')
     expect(screen.queryByText(/especialitat/i)).not.toBeInTheDocument()
+
+    expect(mockHandleFilters).toHaveBeenCalledWith({
+      itinerarySlug: 'react',
+    })
+  })
+
+  it('deletes itinerary deselected and sends other filters to parent', async () => {
+    render(
+      <FiltersWidget
+        filters={{ status: 'ACTIVE' }}
+        handleFilters={mockHandleFilters}
+      />
+    )
+
+    await waitFor(() =>
+      expect(screen.getByText(/especialitat/i)).toBeInTheDocument()
+    )
+
+    const dropdownHeader = screen.getByTestId('dropdown-header')
+
+    fireEvent.click(dropdownHeader)
+
+    const reactOption = screen.getByText('Frontend React')
+    expect(reactOption).toBeInTheDocument()
+
+    fireEvent.click(reactOption)
+
+    expect(dropdownHeader).toHaveTextContent('Frontend React')
+
+    expect(mockHandleFilters).toHaveBeenCalledWith({
+      itinerarySlug: 'react',
+      status: 'ACTIVE',
+    })
+
+    fireEvent.click(dropdownHeader)
+
+    const deselectReact = screen.getByTestId('deselect-1')
+
+    fireEvent.click(deselectReact)
+
+    expect(mockHandleFilters).toHaveBeenCalledWith({
+      status: 'ACTIVE',
+    })
   })
 })