+++ web/usuaris/src/hooks/useGetUsers.tsx
@@ -1,11 +1,13 @@
 import { useQuery } from '@tanstack/react-query'
-import { type TUserData } from '../types'
+import { TFilters, type TUserData } from '../types'
 import { getUsers } from '../helpers'
+import { buildQueryString } from '../helpers/filters'
 
-export const useGetUsers = () => {
-  const { isLoading, isError, data } = useQuery<TUserData[]>({
-    queryKey: ['users'],
-    queryFn: getUsers,
-  })
+export const useGetUsers = (filters: TFilters = {}) => {
+  const { isLoading, isError, data } = useQuery<TUserData[]>(
+    ['users', buildQueryString(filters) || ''],
+    () => getUsers(buildQueryString(filters) || '')
+  )
+
   return { isLoading, isError, data }
 }