+++ web/usuaris/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.11.0] - 2024-04-23
+
+### Added
+
+- Updated Itinerary dropdown with deselect function
+- Added filters to fetch Users
+- Added Itinerary filter feature
+
 ## [0.10.0] - 2024-04-18
 
 ### Added